Reworked downloading Vehicles info
