+++ src/main/java/org/opentripplanner/hasura_client/hasura_mappers/BikeStationsMapper.java
@@ -1,0 +1,12 @@
+package org.opentripplanner.hasura_client.hasura_mappers;
+
+import org.opentripplanner.hasura_client.hasura_objects.BikeStationHasura;
+import org.opentripplanner.routing.bike_rental.BikeRentalStation;
+
+public class BikeStationsMapper extends HasuraToOTPMapper<BikeStationHasura, BikeRentalStation> {
+    @Override
+    protected BikeRentalStation mapSingleHasuraObject(BikeStationHasura hasuraObject) {
+//        TODO miron
+        return null;
+    }
+}
