+++ src/main/java/org/opentripplanner/hasura_client/hasura_objects/HasuraObject.java
@@ -1,0 +1,6 @@
+package org.opentripplanner.hasura_client.hasura_objects;
+
+public class HasuraObject {
+}
+
+
