+++ src/main/java/org/opentripplanner/updater/vehicle_sharing/vehicles_positions/BikeStationsGraphWriterRunnable.java
@@ -1,0 +1,13 @@
+package org.opentripplanner.updater.vehicle_sharing.vehicles_positions;
+
+import org.opentripplanner.routing.graph.Graph;
+import org.opentripplanner.updater.GraphWriterRunnable;
+
+class BikeStationsGraphWriterRunnable implements GraphWriterRunnable {
+
+
+    @Override
+    public void run(Graph graph) {
+//        TODO
+    }
+}
