+++ src/main/java/org/opentripplanner/updater/vehicle_sharing/vehicles_positions/VehicleProvider.java
@@ -1,0 +1,32 @@
+package org.opentripplanner.updater.vehicle_sharing.vehicles_positions;
+
+public class VehicleProvider {
+
+    private int id;
+    private String name;
+    private boolean available;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public boolean isAvailable() {
+        return available;
+    }
+
+    public void setAvailable(boolean available) {
+        this.available = available;
+    }
+}
